backend-agents/pyproject.toml
[project]
name = "backend-agents"
version = "0.1.0"
description = "AWS Strands Agents Service"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[gemini]",
    "strands-agents-tools",
    "mcp",
    "ragie",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "httpx",
    "google-generativeai>=0.8.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff"
]
